# backend/openapi.yaml
openapi: 3.0.0
paths:
...
  /locations/regions:
    get:
      operationId: LocationsController_listRegions
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    region_id:
                      type: string
                    code_region:
                      type: string
                    nom_region:
                      type: string
                      nullable: true
                    code_commune_cheflieu:
                      type: string
                      nullable: true
                    cog_millesime:
                      type: integer
                      nullable: true
                  required:
                    - region_id
                    - code_region
      tags: &a1
        - locations
  /locations/regions/{codeRegion}:
    get:
      operationId: LocationsController_getRegion
      parameters:
        - name: codeRegion
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  region_id:
                    type: string
                  code_region:
                    type: string
                  nom_region:
                    type: string
                    nullable: true
                  code_commune_cheflieu:
                    type: string
                    nullable: true
                  cog_millesime:
                    type: integer
                    nullable: true
                required:
                  - region_id
                  - code_region
        "404":
          description: Unknown region code
      tags: *a1
  /locations/departments:
    get:
      operationId: LocationsController_listDepartments
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 200
        - name: code_region
          required: false
          in: query
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    departement_id:
                      type: string
                    code_departement:
                      type: string
                    nom_departement:
                      type: string
                      nullable: true
                    code_region:
                      type: string
                      nullable: true
                    code_commune_cheflieu:
                      type: string
                      nullable: true
                    cog_millesime:
                      type: integer
                      nullable: true
                  required:
                    - departement_id
                    - code_departement
      tags: *a1
  /locations/departments/{codeDepartement}/communes:
    get:
      operationId: LocationsController_communesByDepartement
      parameters:
        - name: codeDepartement
          required: true
          in: path
          schema:
            type: string
        - name: offset
          required: false
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 500
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    location_id:
                      type: string
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    code_departement:
                      type: string
                      nullable: true
                    code_region:
                      type: string
                      nullable: true
                    type_commune:
                      type: string
                      nullable: true
                    code_commune_parent:
                      type: string
                      nullable: true
                    cog_millesime:
                      type: integer
                      nullable: true
                  required:
                    - location_id
                    - code_commune
      tags: *a1
  /locations/departments/{codeDepartement}:
    get:
      operationId: LocationsController_getDepartment
      parameters:
        - name: codeDepartement
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  departement_id:
                    type: string
                  code_departement:
                    type: string
                  nom_departement:
                    type: string
                    nullable: true
                  code_region:
                    type: string
                    nullable: true
                  code_commune_cheflieu:
                    type: string
                    nullable: true
                  cog_millesime:
                    type: integer
                    nullable: true
                required:
                  - departement_id
                  - code_departement
        "404":
          description: Unknown department code
      tags: *a1
  /locations:
    get:
      operationId: LocationsController_search
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            default: 20
        - name: q
          required: false
          in: query
          description: Search by commune name or INSEE code prefix.
          schema: {}
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    location_id:
                      type: string
                    code_commune:
                      type: string
                    nom_commune:
                      type: string
                      nullable: true
                    code_departement:
                      type: string
                      nullable: true
                    code_region:
                      type: string
                      nullable: true
                    type_commune:
                      type: string
                      nullable: true
                    code_commune_parent:
                      type: string
                      nullable: true
                    cog_millesime:
                      type: integer
                      nullable: true
                  required:
                    - location_id
                    - code_commune
      tags: *a1
  /locations/{codeCommune}:
    get:
      operationId: LocationsController_getByCodeCommune
      parameters:
        - name: codeCommune
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: object
                properties:
                  location_id:
                    type: string
                  code_commune:
                    type: string
                  nom_commune:
                    type: string
                    nullable: true
                  code_departement:
                    type: string
                    nullable: true
                  code_region:
                    type: string
                    nullable: true
                  type_commune:
                    type: string
                    nullable: true
                  code_commune_parent:
                    type: string
                    nullable: true
                  cog_millesime:
                    type: integer
                    nullable: true
                required:
                  - location_id
                  - code_commune
        "404":
          description: Unknown commune code
      tags: *a1
...
info:
  title: Homepedia API
  description: API on top of dbt marts from data_platform.
  version: 1.0.0
  contact: {}
...
servers: []
...